Separate components into their own folder.
