+++ components/asp/decoder/CdlLogLine.py
@@ -1,4 +1,4 @@
-from application.system_processor.decoder.CDL_CONSTANTS import LINE_TYPE_DELIMITER, LINE_TYPE
+from decoder.CDL_CONSTANTS import LINE_TYPE_DELIMITER, LINE_TYPE
 import json
 
 class CdlLogLine:
