+++ components/asp/decoder/Variable.py
