+++ components/query_handler/requirements.txt
@@ -1,0 +1,2 @@
+websockets
+clp-ffi-py
